work on tests

+++ pyHTTPCache.py
@@ -55,24 +55,41 @@ class Resource(object):
 		self.fs_root = '/tmp/pyHTTPCache'
 
 
+	# generate md5 based on url
 	@LazyProperty
 	def md5(self):
+
 		return hashlib.md5(self.url).hexdigest()
 
 
+	# path prop based on cache root and md5 checksum of url
 	@LazyProperty
 	def path(self):
+
 		return "/".join([self.fs_root, self.md5])
 
 
-	def _set_cache(self):
-		r = requests.get(self.url,stream=True)
-		with open(self.path,'w') as fd:
-			for chunk in r.iter_content(chunk_size):
-				fd.write(chunk)
+	# set resource in cache
+	def _set_cache(self, force=False):
+
+		if not self._cache_check() or force:
+			print "CACHE SET"
+			r = requests.get(self.url,stream=True)
+			with open(self.path,'w') as fd:
+				for chunk in r.iter_content(chunk_size):
+					fd.write(chunk)
 
 
+	# remove resource from cache
+	def _del_cache(self, force=False):
+
+		print "CACHE DEL"
+		return os.remove(self.path)
+
+
+	# check cache
 	def _cache_check(self):
+
 		return os.path.exists(self.path)
 
 
@@ -85,8 +102,12 @@ def tests():
 		'image' : Resource(url='http://digital.library.wayne.edu/loris/fedora:wayne:vmc77431_1%7Cvmc77431_1_JP2/full/full/0/default.jpg')
 	}
 
+	# clean test area
+
+
 	for k in test_resources:
-		print "Tests for %s type" % k
+		print "\n\n"
+		print "#### Tests for %s type ####" % k
 
 		# resource example
 		r = test_resources[k]
@@ -99,13 +120,17 @@ def tests():
 		print "setting cache"
 		r._set_cache()
 
-		# set check cache
+		# set re-check cache
 		print "re-check cache..."
 		print r._cache_check()
 
+		# force re-set
+		print "forcing re-set"
+		r._set_cache(force=True)
 
-
-
+		# del cache
+		print "removing from cache"
+		r._del_cache()
